feat(web): add pre-alpha web viewer crate with WASM/WebGPU support
- Add `web/` crate: Rust WASM bridge + TypeScript viewer component
  with REPL, object list, sequence viewer panels
- Fix WGSL shaders for WebGPU: textureSample → textureSampleLevel
  in shadow/AO (uniform control flow), add @interpolate(flat) on
  integer vertex outputs
- Feature-gate crates for WASM: pymol-scene `windowing`, pymol-render
  `parallel` (rayon), pymol-framework `dirs`, pymol-io drop unused
  memmap2
- Add silhouette post-process pass to web renderer
- Add desktop-style panel layout and minimal viewer example
- Add `build-web` job to release workflow
- Bump all versions to 0.2.2

+++ crates/pymol-select/src/result.rs
@@ -13,21 +13,21 @@ use pymol_mol::AtomIndex;
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct SelectionResult {
     /// Bitset where bit i is set if atom i is selected
-    bits: BitVec<u64, Lsb0>,
+    bits: BitVec<usize, Lsb0>,
 }
 
 impl SelectionResult {
     /// Create a new empty selection (no atoms selected)
     pub fn new(atom_count: usize) -> Self {
         SelectionResult {
-            bits: bitvec![u64, Lsb0; 0; atom_count],
+            bits: bitvec![usize, Lsb0; 0; atom_count],
         }
     }
 
     /// Create a selection with all atoms selected
     pub fn all(atom_count: usize) -> Self {
         SelectionResult {
-            bits: bitvec![u64, Lsb0; 1; atom_count],
+            bits: bitvec![usize, Lsb0; 1; atom_count],
         }
     }
